+++ changelog.md
@@ -1,3 +1,7 @@
+### version 0.6.0
+Moving the Metadata Reader & Writer into bundled plugins automatically installed in parallel with ePub Extended Metadata
+(following a discussion with Kovid about the better implementation for "multiples plugin")
+
 ### version 0.5.0
 First release
 Beta public test